# docs/swagger.yaml
basePath: /
definitions:
  api.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.Ok:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  api.RegisterRequest:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - firstname
    - lastname
    - password
    - username
    type: object
  domain.Appointment:
    properties:
      doctor_id:
        type: integer
      id:
        type: integer
      patient_id:
        type: integer
      time:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: 211022@astanait.edu.kz
    name: Bakhityar
  description: Hospital application where patients and doctors can decide on appointment
    time
  title: Hospital Application
  version: 0.0.1
paths:
  /user/appointment/create:
    post:
      consumes:
      - application/json
      description: Authorized user can create appointments
      parameters:
      - description: req body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/domain.Appointment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Ok'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: User creates a new appointment
      tags:
      - appointment
  /user/login:
    post:
      consumes:
      - application/json
      description: After successfully identifying oneself, the user is provided with
        AccessToken for authorization purposes
      parameters:
      - description: req body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Ok'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: User logs in
      tags:
      - auth
  /user/register:
    post:
      consumes:
      - application/json
      description: Create new user
      parameters:
      - description: req body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create user
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"